test: split VarsUnitTest into per-subject test files
Replace the monolithic VarsUnitTest with one file per subject:
BranchPolicyTest, SemVerTest, ConventionalCommitTest (pure value
classes, no JPU), and RequireEnvTest, CaptureBuildInfoTest, TagBuildTest,
EventuallyTest, WithParallelMatrixTest, NotifyBuildTest,
RequireConventionalPrTitleTest (all extend PipelineTest abstract base).
